Update v1.5
Code optimization
More simplified installation
autodetection API Request and Responses need to specify the path
autodetection Pathfile need to specify the path
autodetection of the return URL and cancellation (no need to indicate)
autodetection ipn url of atos (no need to indicate)
Creating a new template for displaying ATOS payment options, a
completely independent in your template (to be uploaded in the templates
/ directory frontoffice)
Using the failed-page order in case of refusal or cancellation
$ {Variable} refusmotif available to display the reason for the refusal
or cancellation to indicate that just himself.
Using the apage-order Placed in the context of acceptance of payment.
Version 1.5